realtime-voice-agent-system/requirements.txt
openai==1.35.3
httpx==0.27.0
python-dotenv==1.0.1
tiktoken==0.7.0
sounddevice==0.4.7
numpy==1.26.4
scipy==1.13.1
kokoro==0.9.4
torch==2.3.1
transformers==4.41.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pytest==8.2.2
pytest-cov==5.0.0
pytest-asyncio==0.23.7
hypothesis==6.104.2
black==24.4.2
flake8==7.1.0

